OhdsiSharing-package {OhdsiSharing}	R Documentation

OhdsiSharing: Package for sharing of the results of the OHDSI tools

Description

Package for sharing of the results of the OHDSI tools, with functions for encrypting results and sending results through SFTP to a central site.

Author(s)

Maintainer: Martijn Schuemie schuemie@ohdsi.org

Authors:

Marc Suchard

Lee Evans

See Also

Useful links:

https://ohdsi.github.io/OhdsiSharing

https://github.com/OHDSI/OhdsiSharing

Report bugs at https://github.com/OHDSI/OhdsiSharing/issues

[Package OhdsiSharing version 0.2.2 Index]