compressFolder {OhdsiSharing}	R Documentation

Compress a folder

Description

Compress a folder

Usage

compressFolder(sourceFolder, targetFileName)

Arguments

sourceFolder
Name of the folder that must be compressed.

targetFileName
Name of the file that will hold the compressed data.

Details

Compresses all files in a folder and its subfolders, and stores it in a single zip file.

Examples

## Not run: 
# Create a folder with some data
dir.create("test")
data <- data.frame(x = runif(1000), y = 1:1000)
saveRDS(data, "test/data1.rds")
saveRDS(data, "test/data2.rds")

compressFolder("test", "data.zip")
decompressFolder("data.zip", "test2")

## End(Not run)

[Package OhdsiSharing version 0.2.2 Index]