decryptFile {OhdsiSharing}	R Documentation

Decrypt a data file

Description

Decrypt a data file

Usage

decryptFile(sourceFileName, targetFileName, privateKeyFileName)

Arguments

sourceFileName
Name of the file that must be decrypted.

targetFileName
Name of the file that will hold the unencrypted data.

privateKeyFileName
Name of the file where the private key is stored.

Details

Decrypts the data using the provided private key.

Examples

## Not run: 
generateKeyPair("public.key", "private.key")
data <- data.frame(x = runif(1000), y = 1:1000)
saveRDS(data, "data.rds")
encryptFile("data.rds", "data.rds.enc", "public.key")
decryptFile("data.rds.enc", "data2.rds", "private.key")

## End(Not run)

[Package OhdsiSharing version 0.2.2 Index]