encryptFile {OhdsiSharing}	R Documentation

Encrypt a data file

Description

Encrypt a data file

Usage

encryptFile(sourceFileName, targetFileName, publicKeyFileName)

Arguments

sourceFileName
Name of the file that must be encrypted.

targetFileName
Name of the file that will hold the encrypted data.

publicKeyFileName
Name of the file where the public key is stored.

Details

Encrypts the data using the provided public key.

Examples

## Not run: 
generateKeyPair("public.key", "private.key")
data <- data.frame(x = runif(1000), y = 1:1000)
saveRDS(data, "data.rds")
encryptFile("data.rds", "data.rds.enc", "public.key")

## End(Not run)

[Package OhdsiSharing version 0.2.2 Index]