generateKeyPair {OhdsiSharing}	R Documentation

Create a public-private key pair

Description

Create a public-private key pair

Usage

generateKeyPair(publicKeyFileName, privateKeyFileName)

Arguments

publicKeyFileName
Name of the file where the public key should be stored.

privateKeyFileName
Name of the file where the private key should be stored.

Details

Creates an RSA 4096-bit public-private key pair. The public key can be used to encrypt data, and only with the private key can the data be decrypted.

Examples

## Not run: 
generateKeyPair("public.key", "private.key")

## End(Not run)

[Package OhdsiSharing version 0.2.2 Index]