sftpGetFiles {OhdsiSharing}	R Documentation

Get one or more files from the SFTP server

Description

Get one or more files from the SFTP server

Usage

sftpGetFiles(
  sftpConnection,
  remoteFileNames,
  localFolder = getwd(),
  localFileNames = file.path(localFolder, remoteFileNames)
)

Arguments

sftpConnection
An SftpConnection object as created by the sftpConnect function.

remoteFileNames
The name of the file(s) to get from the server.

localFolder
The path of a local folder where all files will be stored. Is ignored if localFileNames is provided.

localFileNames
The name the file(s) should have locally. If not provided, the files will be given the same names as on the server.

[Package OhdsiSharing version 0.2.2 Index]